+++ src/action.ts
@@ -3,14 +3,14 @@ import * as fs from "fs";
 import * as path from "path";
 import * as yaml from "js-yaml";
 import * as deepl from "deepl-node";
-import { readL10nConfig } from "./utils";
+import { extractKeyNameFromText, readL10nConfig } from "./utils";
 
 let options = {
   // flutter options
   arbDirName: "lib/l10n",
   templateArbFile: "app_en.arb",
   // package specific options
-  mainLocaleCode: "en",
+  // mainLocaleCode: "en",
   autoTranslate: true,
   keyPrefix: "AppLocalizations.of(context).",
   importStr: "",
@@ -26,7 +26,7 @@ function setupConfig() {
     ...options, // Preserve existing defaults
     arbDirName: l10nConfig["arb-dir"] ?? options.arbDirName,
     templateArbFile: l10nConfig["template-arb-file"] ?? options.templateArbFile,
-    mainLocaleCode: l10nConfig["main-locale"] ?? options.mainLocaleCode,
+    // mainLocaleCode: l10nConfig["main-locale"] ?? options.mainLocaleCode,
     autoTranslate: l10nConfig["translate"] ?? options.autoTranslate,
     importStr: l10nConfig["import-line"] ?? options.importStr,
     keyPrefix: l10nConfig["key-prefix"] ?? options.keyPrefix,
@@ -45,10 +45,13 @@ export async function extractToArb(
   const editor = vscode.window.activeTextEditor;
   if (!editor) return;
 
-  const key = await promptForKey();
-  if (!key) return;
+  const value = text.slice(1, -1); // Remove quotes from the string literal : "text" -> text
 
-  const value = text.slice(1, -1); // Remove quotes from the string literal
+  const key = options.autoGenerateKeyName
+    ? extractKeyNameFromText(value)
+    : await promptForKey();
+
+  if (!key) return;
 
   const arbWriteSucces = updateArbFiles(key, value);
   if (!arbWriteSucces) return;
@@ -106,7 +109,7 @@ export function getDeeplApiKey(): string | undefined {
 
 async function translateText(
   text: string,
-  sourceLang: string,
+  // sourceLang: string,
   targetLang: string
 ): Promise<string> {
   const authKey = getDeeplApiKey();
@@ -116,11 +119,18 @@ async function translateText(
   }
   const translator = new deepl.Translator(authKey);
 
+  // https://developers.deepl.com/docs/resources/supported-languages
+  targetLang = targetLang.toLowerCase();
+  if (targetLang === "en") targetLang = "en-US";
+  if (targetLang === "pt") targetLang = "pt-PT";
+  if (targetLang === "zh") targetLang = "zh-HANS";
+
   const result = await translator.translateText(
     text,
-    sourceLang as deepl.SourceLanguageCode,
+    null, // sourceLang as deepl.SourceLanguageCode,
     targetLang as deepl.TargetLanguageCode
   );
+
   return result.text;
 }
 
@@ -130,10 +140,10 @@ async function updateArbFiles(key: string, value: string): Promise<boolean> {
   const arbDirPath = path.join(workspacePath, options.arbDirName);
 
   // app_en.arb -> app_fr.arb
-  const mainArbFileName = options.templateArbFile.replace(
-    "en.arb",
-    options.mainLocaleCode + ".arb"
-  );
+  // const mainArbFileName = options.templateArbFile.replace(
+  //   "en.arb",
+  //   options.mainLocaleCode + ".arb"
+  // );
 
   let ok = true;
 
@@ -142,11 +152,11 @@ async function updateArbFiles(key: string, value: string): Promise<boolean> {
   for (const arbFile of arbFiles) {
     const fullArbPath = path.join(arbDirPath, arbFile);
     const val =
-      options.autoTranslate && arbFile !== mainArbFileName
+      options.autoTranslate && arbFile //!== mainArbFileName
         ? await translateText(
             value,
-            options.mainLocaleCode,
-            arbFile.split("_")[1].split(".")[0]
+            // options.mainLocaleCode,
+            arbFile.split("_")[1].split(".")[0] // lang_en.arb -> en
           )
         : value;