+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2.0.0
+- simiplified the options & defaults values
+- added a new option to auto generate keys 
+- added a new option to import line if needed
+- update docs
+
 ## 1.0.0
 
 Initial release !